Fix build with release candidate version of shortened version numbers.
We have now one fewer number in the version numbers. This means that the
release candidate identifier "rc" is in the third position.

When git.rc is compiled on Windows, we pick the third position of the
version number as one of the arguments of the FILEVERSION statement.
However, all these arguments must be numeric, but "rc0" is not. To make
all arguments numeric, special-case release candidates: use the previous
version number, but with 90 + rc-number.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>